Hotel Service (#11)
* Added HotelService

* Added Hotels database initialization

* Added hotels.csv

* Added loading hotels from csv

* Updated hotels.csv

* Added Dockerfile

+++ backend/TripBooker.TransportService/Model/ReservationModel.cs
@@ -1,5 +1,4 @@
 ﻿using TripBooker.Common;
-using TripBooker.Common.Transport;
 
 namespace TripBooker.TransportService.Model;
 
@@ -9,7 +8,7 @@ internal class ReservationModel : EventModel
 
     public int Places { get; set; }
 
-    public int Price { get; set; }
+    public double Price { get; set; }
 
     public ReservationStatus Status { get; set; }
 }
